tentative new name for the Docker image, updated instructions

+++ buildStartupKits.sh
@@ -1,6 +1,10 @@
 #!/usr/bin/env bash
 
-DOCKER_IMAGE=nvflare-pt-dev:odelia
+if [ "$#" -ne 1 ]; then
+    echo "Usage: buildStartupKits.sh SWARM_PROJECT.yml"
+    exit 1
+fi
 
-# docker build -t $DOCKER_IMAGE . -f docker_config/Dockerfile_nfcore
+DOCKER_IMAGE=jefftud/odelia:1.0rc1
+
 docker run --rm -it -u $(id -u):$(id -g) -v /etc/passwd:/etc/passwd -v /etc/group:/etc/group -v ./:/workspace/ -w /workspace/ $DOCKER_IMAGE /bin/bash -c "nvflare provision -p $1"
